Move run attrs tests to their own file
